Rename vds package to core
