+++ internal/core/attribute.cpp
