+++ internal/core/attribute.hpp
