+++ internal/core/axis.cpp
