+++ internal/core/boundingbox.cpp
