+++ internal/core/boundingbox.hpp
