+++ internal/core/capi.cpp
