+++ internal/core/capi.h
