+++ internal/core/connection.go
@@ -1,4 +1,4 @@
-package vds
+package core
 
 import (
 	"context"
