+++ internal/core/cppapi.hpp
