+++ internal/core/cppapi_data.cpp
