+++ internal/core/cppapi_metadata.cpp
