+++ internal/core/direction.cpp
