+++ internal/core/direction.hpp
