+++ internal/core/exceptions.hpp
