+++ internal/core/metadatahandle.cpp
