+++ internal/core/metadatahandle.hpp
