+++ internal/core/vds_test.go
@@ -1,4 +1,4 @@
-package vds
+package core
 
 import (
 	"encoding/binary"
