+++ internal/core/verticalwindow.cpp
