+++ internal/core/verticalwindow.hpp
